# web/storage/api-docs/api-docs.yaml
openapi: 3.0.0
info:
  title: 'Payment Gateways API Microservice'
  description: 'Payment Gateways API Microservice, Version v1'
  contact:
    name: 'Support Team'
    email: support@ultainfinity.com
  version: v1
servers:
  -
    url: 'http://localhost:8100/v1'
    description: 'Payment Gateways API Microservice, Version v1'
paths:
...
  /payments/charge:
    post:
      tags:
        - Payments
      summary: 'Make payment and charge wallet balance or invoice'
      description: 'Make payment and charge wallet balance or invoice'
      operationId: 576db5189fef552c041f1696c0303952
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                gateway:
                  description: 'Payment gateway'
                  type: string
                  default: bitpay
                amount:
                  description: 'The amount of money replenished to the balance'
                  type: integer
                  default: 1000
                currency:
                  description: 'Currency of balance'
                  type: string
                  default: GBP
                service:
                  description: 'Target service: '
                  type: string
              type: object
      responses:
        '200':
          description: Success
      security:
        -
          default:
            - ManagerRead
            - User
            - ManagerWrite
      x-auth-type: 'Application & Application User'
      x-throttling-tier: Unlimited
      x-wso2-application-security:
        security-types:
          - oauth2
        optional: 'false'
  '/payments/{id}':
    get:
      tags:
        - Payments
      summary: 'Get detail info about payment transaction'
      description: 'Get detail info about payment transaction'
      operationId: f544d55b7f000c8f444cc28872a25fc6
      parameters:
        -
          name: id
          in: path
          description: 'Payment transaction ID'
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 'Data of payment transaction'
        '404':
          description: 'Payment transaction not found'
      security:
        -
          default:
            - ManagerRead
            - User
            - ManagerWrite
      x-auth-type: 'Application & Application User'
      x-throttling-tier: Unlimited
      x-wso2-application-security:
        security-types:
          - oauth2
        optional: 'false'
...
components:
  securitySchemes:
    default:
      type: oauth2
      description: 'Auth Scheme'
      flows:
        implicit:
          authorizationUrl: 'https://sumraid.com/oauth2'
          scopes:
            ManagerRead: 'Manager can read'
            User: 'User access'
            ManagerWrite: 'Manager can write'